Move jsx from utils to component / make css file
