+++ src/components/Folder/Banner/Profile/styles.js
@@ -1,0 +1,34 @@
+import { css } from "@emotion/react";
+
+export const divStyle = css`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  background: var(--Linkbrary-bg);
+`;
+
+export const imgWrapperStyle = css`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 60px;
+  height: 60px;
+  margin-bottom: 12px;
+`;
+export const imgStyle = css`
+  width: 100%;
+  height: 100%;
+`;
+export const pStyle = css`
+  margin: 0;
+  font-size: 16px;
+  line-height: 24px;
+`;
+export const h1Style = css`
+  font-size: 40px;
+  font-weight: 600;
+  line-height: 48px;
+  margin-top: 20px;
+  margin-bottom: 60px;
+`;
