+++ src/components/Layout/index.jsx
@@ -1,4 +1,4 @@
-import {React, Fragment} from "react";
+import { React, Fragment } from "react";
 import Header from "./Header";
 import Footer from "./Footer";
